Make timeouts more flexible
Limit more actions with timeout. Let timeouts decrease if several
events time out, slowing down crawling. This is a hack which is
designed mainly to solve problem of endless animations on page,
which handgs settle tracker. A more gentle solution would be to
detect such elements that animate endlessly and ignore them, but
this is harder to implement